requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
pillow-avif-plugin>=1.4.0
orjson>=3.10.0
passlib==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
